+++ Ladybird/Android/src/main/cpp/WebContentService.h
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2023, Andrew Kaster <akaster@serenityos.org>
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#pragma once
+
+#include <AK/Error.h>
+
+ErrorOr<int> web_content_main(int ipc_socket, int fd_passing_socket);
